LayoutTests/media/audio-playback-restriction-play-expected.txt

Test that, when RequireUserGestureForAudioRateChange is set, starting playback of an audio file requires a user gesture.


RUN(internals.setMediaElementRestrictions(video, "RequireUserGestureForAudioRateChange"))
RUN(video.src = findMediaFile('audio', 'content/test'))
EVENT(canplaythrough)
RUN(video.play())
EXPECTED (video.paused == 'true') OK
Running with keyDown.
RUN(video.play())
EXPECTED (video.paused == 'false') OK
END OF TEST
